refactor(mark): merge api components into mark module and update references
- Relocate `MarkDto`, `MarkOccurrenceDto`, and `MarkUpdateDto` from `mark-api` to `mark` module
- Move `MarkService` from `mark-api` to `mark` module and update package to `pt.estga.mark.services`
- Update imports in `MarkController` and `MarkServiceImpl` to reflect the new `MarkService` location
- Remove `spring-data-commons` dependency from `mark-api` module's `pom.xml`
